Extract common focus management logic into shared hook
Created useModalFocusManagement hooks to reduce code duplication:
- useModalFocusManagement: For modals that conditionally render when open
- useModalFocusManagementUnmounting: For modals that unmount when closed

All three modal components now use these shared hooks instead of
duplicating the same focus management logic.

Benefits:
- DRY: Common code in one place
- Consistency: Same behavior across all modals
- Maintainability: Single source of truth for focus management
